fix: extract override_during attribute to own file and move mock_resource/mock_data patching into mock_provider_block
